+++ client/shared/Hooking.Patterns.cpp
@@ -158,7 +158,7 @@ bool pattern::ConsiderMatch(uintptr_t offset)
 
 	char* ptr = reinterpret_cast<char*>(offset);
 
-	for (int i = 0; i < m_size; i++)
+	for (size_t i = 0; i < m_size; i++)
 	{
 		if (mask[i] == '?')
 		{
